+++ client/package.json
@@ -43,6 +43,8 @@
     "vite": "^6.0.0"
   },
   "dependencies": {
+    "@reown/appkit": "^1.6.0",
+    "@reown/appkit-adapter-solana": "^1.6.0",
     "@solana/wallet-adapter-base": "^0.9.23",
     "@solana/wallet-adapter-wallets": "^0.19.32",
     "@solana/web3.js": "^1.95.8",
@@ -50,6 +52,7 @@
     "@svelte-on-solana/wallet-adapter-ui": "^1.0.21-alpha.0",
     "@tailwindcss/container-queries": "^0.1.1",
     "@tailwindcss/typography": "^0.5.15",
+    "react": "^19.0.0",
     "tailwindcss-animate": "^1.0.7"
   }
 }
